将navbar改为vue3格式
