+++ src/router/index.js
@@ -4,7 +4,7 @@ import Couple from '../views/Couple.vue'
 import Login from '../views/Login.vue'
 import Register from '../views/Register.vue'
 import Operation from '../views/Operation.vue'
-import OperationThree from '../views/OperationThree.vue'
+// import OperationThree from '../views/OperationThree.vue'
 const routes = [
   {
     path: '/',
@@ -17,11 +17,11 @@ const routes = [
     name: 'Operation',
     component: Operation,
   },
-  {
-    path: '/operationthree',
-    name: 'OperationThree',
-    component: OperationThree,
-  },
+  // {
+  //   path: '/operationthree',
+  //   name: 'OperationThree',
+  //   component: OperationThree,
+  // },
   {
     path: '/couple',
     name: 'Couple',
