+++ src/views/Operation.vue
@@ -74,7 +74,7 @@
           <p v-show="!isShow">暂无运行结果...</p>
           <div v-show="isShow" >
             <!-- <el-card v-for="(item, index) in modelConfig" :key="index" style="margin:2px">             -->
-            <el-form :label-position="labelPosition">                  
+            <el-form :label-position="labelPosition">
               <el-form-item>
                 <p style="float: left;margin-top: 35px;margin-right: 20px;margin-left: 20px;">输出结果：</p>
                 <div style="text-align:center;vertical-align:middle;margin-top: 25px;">
@@ -102,11 +102,12 @@
           type="primary"
           plain
           style="margin:10px;position:relative;left: 40%;top: 20px;"
+          @click="showMap"
           >
-          <router-link to="couple">
+          <!-- <router-link to="couple"> -->
             <el-icon style="margin-right:5px"><promotion /></el-icon>
             耦合结果展示
-          </router-link>
+          <!-- </router-link> -->
         </el-button>
          
       </el-card>
@@ -196,364 +197,301 @@
     </el-dialog>
   </div>
 </template>
-<script>
+<script setup>
 // import vkbeautify from "vkbeautify";
 import Navbar  from '../components/Navbar'
-import { UploadFilled,Bottom,Pointer,View,
-VideoPlay,Promotion } from '@element-plus/icons-vue'
-import { ref } from 'vue';
-import { fa } from 'element-plus/lib/locale';
-export default {
-  name:"Operation",
-  components:{
-    Navbar,
-    UploadFilled,
-    Bottom,
-    Pointer,
-    View,
-    VideoPlay,
-    Promotion,
-    ref,
+import { UploadFilled,Bottom,Pointer,View,VideoPlay,Promotion } from '@element-plus/icons-vue'
+import { reactive, ref, computed ,toRefs} from 'vue';
+import axios from 'axios'
+import {ElMessage } from 'element-plus'
+import {  useRouter} from 'vue-router'
+const router = useRouter();
+let dialog = reactive({
+    dialogVisible:false,
+    deployDocument: false,          
+});
+
+let {dialogVisible, deployDocument} = toRefs(dialog);
+
+let load = reactive({
+    loading: false,
+    downloadLoading:false,
+    downloadFileLoading:false,
+});
+let {loading, downloadLoading,downloadFileLoading} = toRefs(load);
+
+let modelData = reactive({
+  modelList: [],
+  modelConfig: [],
+  loadModel: {},
+  loadModels:[],
+  uploadFiles:[],
+})
+
+let {modelList, modelConfig,loadModel,loadModels,uploadFiles} = toRefs(modelData);
+let configUpload = ref();
+let value = ref('');
+let xml_show = ref("");
+let isShow = ref(false);
+let labelPosition = ref('left');
+let outputPath = ref('E:\\research\\yangtze_delta');
+
+let list = [
+  {
+    oid: "3f6857ba-c2d2-4e27-b220-6e5367803a12",
+    name: "SWMM",
+    description:
+      "SWMM在1971年由美国环境保护局联合多家企业单位共同开发问世，最初只能完成一些诸如不确定性参数验证以及简单的雨洪过程模拟，后来经历多次的升级迭代，被广泛应用于城市雨水径流模拟、地下管网的水流模拟、LID模拟、雨水模拟以及与其他的排水系统相结合进行各种排水规划、设计与排水分析工作。目前版本已更新到SWMM5，可以对研究区域的输入数据进行编辑，以及可以执行各种水文、水力以及水质的模拟工作。",
+    author: "美国环境保护局",
+    select:false,
+    output:'output_swmm',
   },
-  data() {
-    return {
-      dialogVisible: false,
-      deployDocument: false,
-      list: [
-        {
-          oid: "3f6857ba-c2d2-4e27-b220-6e5367803a12",
-          name: "SWMM",
-          description:
-            "SWMM在1971年由美国环境保护局联合多家企业单位共同开发问世，最初只能完成一些诸如不确定性参数验证以及简单的雨洪过程模拟，后来经历多次的升级迭代，被广泛应用于城市雨水径流模拟、地下管网的水流模拟、LID模拟、雨水模拟以及与其他的排水系统相结合进行各种排水规划、设计与排水分析工作。目前版本已更新到SWMM5，可以对研究区域的输入数据进行编辑，以及可以执行各种水文、水力以及水质的模拟工作。",
-          author: "美国环境保护局",
-          select:false,
-          output:'output_swmm',
-        },
-        {
-          oid: "5738ef7c-a5ac-46b5-a347-3c823f71b3a7",
-          name: "LISFLOOD-FP",
-          description:
-            "LISFLOOD-FP是一种二维水动力模型，由英国布里斯托尔大学开发并由Paul Bates和Ad De Roo等人在2000年提出，经过多次的更新迭代，目前已广泛应用于一维河道水力模拟以及二维蓄洪区的水力变化，对城市内涝、河道洪泛的研究模拟具有很好的模拟效果。",
-          author: "Paul Bates, Mark Trigg, Jeff Neal and Amy Dabrowa",
-          select:false,
-          output:'output_lisflood_fp',
-        },
-        {
-          oid: "b3a88af1-3568-4219-99c4-aa8fbce3227d",
-          name: "ANUGA",
-          description:
-            "ANUGA模型是一种流体力学建模工具，允许用户在复杂的二维几何形状中建模模拟现实的流动问题。是由Stephen Roberts，Ole Nielsen,Duncan Gray,Jane Sexton等人开发。ANUGA的核心是流体动力学模块，称为Shallow_water，它是基于求解浅水波方程的有限体积方法。研究区域用三角形网格表示。通过求解每个单元内的控制方程，可以跟踪水深和水平动量随时间的变化。",
-          author: "Stephen Roberts，Ole Nielsen,Duncan Gray,Jane Sexton",
-          select:false,
-          output:'output_anuga',
-        }
-      ],
-      modelList: [],
-      modelConfig: [],
-      xml_show: "",
-      loadModel: {},
-      loadModels:[],
-      loading: false,
-      downloadLoading:false,
-      downloadFileLoading:false,
-      isShow:false,
-      labelPosition:'left',
-      output:[{
-          label:'1',
-          path:'',
-          value:'report1.rpt'
-        },{
-          label:'2',
-          path:'',
-          value:'output1.out'
-        },{
-          label:'3',
-          path:'res_clip_testNoBciAndDoubleCouple1',
-          value:'fenhu-0000.wd',
-        },{
-          label:'4',
-          path:'res_clip_testNoBciAndDoubleCouple1',
-          value:'fenhu-0001.wd',
-        },{
-          label:'5',
-          path:'res_clip_testNoBciAndDoubleCouple1',
-          value:'fenhu-0002.wd',
-        },
-      ],
-      modelSelectStatus:'选择',
-      uploadFiles:[],
-      value:ref(''),
-      outputPath:'E:\\research\\yangtze_delta',
-    };
+  {
+    oid: "5738ef7c-a5ac-46b5-a347-3c823f71b3a7",
+    name: "LISFLOOD-FP",
+    description:
+      "LISFLOOD-FP是一种二维水动力模型，由英国布里斯托尔大学开发并由Paul Bates和Ad De Roo等人在2000年提出，经过多次的更新迭代，目前已广泛应用于一维河道水力模拟以及二维蓄洪区的水力变化，对城市内涝、河道洪泛的研究模拟具有很好的模拟效果。",
+    author: "Paul Bates, Mark Trigg, Jeff Neal and Amy Dabrowa",
+    select:false,
+    output:'output_lisflood_fp',
   },
-  methods: {
-    //关闭dialogVisible对话框
-    handleClose(done) {
-      this.dialogVisible = false;
-    },
-    //关闭deployDocument对话框
-    handleCloseDeployDocument(done) {
-      this.deployDocument = false;
-    },
-    //搜索模型
-    searchModelItem() {},
-    //选择需要耦合的模型
-    selectModelItem(index, info) {
-      let modelObj = {
-        oid:info.oid,
-        name:info.name,
-        description:info.description,
-        author:info.author,
-      }
-      this.modelList.push(modelObj);
-      this.$message({
-        message:'选择模型成功',
-        type:'success',
-      })
-      info.select = !info.select;
-    },
-    //载入已选的模型
-    loadModelItem(index, info) {
-      this.xml_show = '';
-      this.fileList = '';
-      this.deployDocument = true;
-      let modelObj = {
-        oid:info.oid,
-        name:info.name,
-        description:info.description,
-        author:info.author,
-      }
-      this.loadModel = modelObj;
-      this.loadModels.push(modelObj);
-    },
-    //调用上传的删除方法，清除上传文件列表
-    clearFiles(){
-      if(this.$refs['configUpload'] !== undefined){        
-        this.$refs['configUpload'].clearFiles();
-      }
-    },
-    //上传文件，并展示内容到页面上
-    uploadChange(file, fileList) {
-      if (file.status !== 'ready') return;
-      let fileObj = {
-        file:fileList,
-        name:this.loadModel.name,
-        description:this.loadModel.description,
-        author:this.loadModel.author,        
-        oid:this.loadModel.oid,
-      }
-      this.uploadFiles.push(fileObj);
-      
-      let _this = this;
-      let form = new FormData();
-      for (let i = 0; i < this.uploadFiles.length; i++) {
-        //上传文件信息以及模型信息
-        form.append("datafile", this.uploadFiles[i].raw);
-      }
-
-      for(let item of this.uploadFiles){
-        //上传文件信息以及模型信息
-        form.append("datafile", item.file[0].raw);
-        form.append("name", item.name);
-        form.append("description", item.description);
-        form.append("author", item.author);
-        form.append("oid", item.oid);
-      }
-      this.axios.post("/api/coupleDocument", form).then(res => {
-        //解析该文件，返回xml字符串，赋值给xml_show
-        if (res.status === 200) {
-          let result = res.data;
-          if (result.code === 0) {
-            this.$message({
-              message: "载入成功",
-              type: "success"
-            });
-            _this.xml_show = result.data;
-          } else {
-            this.$message.error("载入失败");
-          }
-        } else {
-          this.$message.error("服务器error");
-        }
-      });
-    },
-    //载入配置文件到模型模块化耦合控制区操作面板
-    confirmLoad() {    
-      this.modelConfig.push({
-        name: this.loadModel.name,
-        oid:this.loadModel.oid,
-        configName: this.loadModel.name + ".conf",
-      });
-      this.deployDocument = false;
-    },
-    //调用函数
-    invoke() {
-      //根据文件oid，调用exe，生成最终结果，并可视化
-      let form = new FormData();
-      let oid = '', oids = [];
-      for(let item of this.loadModels){
-        if(oid !== item.oid){
-          oid = item.oid;
-          oids.push(oid);
-        }  
-      }
-      if(oids.length < 2){
-        this.$message({
-          message:"模型数目不够2个，无法耦合",
-          type:'error',
-        })        
-      }else{
-        form.append('oids', oids);
-        this.loading = true;
-        let _this = this;
-        //调用invoke接口
-        // this.axios.post('/api/invoke', form).then(res => {
-        //   if(res.status === 200){
-        //     let result = res.data;
-        //     if(result.code === 0){
-        //       this.$message({
-        //         message:'调用耦合方法成功！',
-        //         type:'success',
-        //       })
-        //       _this.outputPath = result.data;
-        //       _this.isShow = true;
-        //     }else{
-        //       this.$message.error(`调用耦合方法失败！,失败原因为：${result.message}`);
-        //     }            
-        //     _this.loading = false;
-        //   }else{
-        //     this.$message.error('服务器error');
-        //   }
-        // })
-        setTimeout(() => {
-          _this.isShow = true;
-          _this.loading = false;
-        }, 1000)
-      }
-    },
-    //配置文件下载函数
-    downloadConfig(oid){
-      console.log(`oid 为 ${oid}`);
-      if(oid === undefined){
-        this.$message({message:'该配置文件无oid，请检查！', type:'error'});
-      }
-      let form = new FormData();
-      form.append('oid',oid);
-      this.downloadLoading = true;
-      // let _this = this;
-      
-      window.location.href = `/api/downloadConfig?oid=${oid}`;  
-      this.downloadLoading = false;
-
-      // this.axios.get('/api/downloadConfig', {
-      //   params:{oid:oid},
-      //   responseType:'blob',
-      // }).then(res => {
-      //     _this.downloadLoading = false;
-      //     if(res.status === 200){
-      //     const content = res.data;
-      //     const blob = new Blob([content]);
-      //     if('download' in document.createElement('a')){
-      //       //非IE下载
-      //       const elink = document.createElement('a');
-      //       elink.download = _this.selected;
-      //       elink.style.display = 'none';
-      //       elink.href = window.URL.createObjectURL(blob);
-      //       document.body.appendChild(elink);
-      //       elink.click();
-      //       window.URL.revokeObjectURL(elink.href);
-      //       document.body.removeChild(elink);
-      //     }else{
-      //       //IE10+下载
-      //       if(typeof window.navigator.msSaveBlob !== 'undefined'){
-      //         window.navigator.msSaveBlob(blob, _this.selected);
-      //       }else{
-      //         let URL = window.URL || window.webkitURL;
-      //         let downloadUrl = URL.createObjectURL(blob);
-      //         window.location = downloadUrl;
-      //       }
-      //     }
-
-      //     let result = res.data;
-      //     if(result.code === -1){
-      //       this.message({
-      //         message:`下载配置文件失败,失败原因为：${result.message}`,
-      //         type:'error',
-      //       })
-      //       console.log(`下载配置文件失败,失败原因为：${result.message}`)
-      //     }//如果不为-1则成功下载，不需要弹框，直接下载即可
-      //   }else{
-      //     this.message({
-      //       message:'服务器相应失败',
-      //       type:'error',
-      //     })
-      //   }
-      // })
-    },
-    //下载输出数据
-    downloadOutput(){
-      if(this.selected === ''){
-        this.$message({message:"未选择需要下载的文件，请先选择待下载文件！", type:"error"})
-        return;
-      }
-      //拼写出下载的文件路径
-      let path = '';
-      for(let item of this.output){
-        if(item.value === this.selected){
-          if(item.path === ''){
-            path = this.outputPath + '/' + this.selected;
-          }else{
-            path = this.outputPath + '/' + item.path + '/' + this.selected;
-          }          
-          break;
-        }
-      }
-      let form = new FormData();
-      form.append('filePath', path);
-      this.downloadFileLoading = true;
-      let _this = this;
-      //利用blob实现axios下载,超时时间五分钟
-      this.axios.post('/api/downloadFile', form, {responseType:'blob'},{timeout:30000}).then(res => {
-        _this.downloadFileLoading = false;
-
-        const content = res.data;
-        const blob = new Blob([content]);
-        if('download' in document.createElement('a')){
-          //非IE下载
-          const elink = document.createElement('a');
-          elink.download = _this.selected;
-          elink.style.display = 'none';
-          elink.href = window.URL.createObjectURL(blob);
-          document.body.appendChild(elink);
-          elink.click();
-          window.URL.revokeObjectURL(elink.href);
-          document.body.removeChild(elink);
-        }else{
-          //IE10+下载
-          if(typeof window.navigator.msSaveBlob !== 'undefined'){
-            window.navigator.msSaveBlob(blob, _this.selected);
-          }else{
-            let URL = window.URL || window.webkitURL;
-            let downloadUrl = URL.createObjectURL(blob);
-            window.location = downloadUrl;
-          }
-        }
-      }).catch(err => {
-        console.log(err);
-      })      
-    }
+  {
+    oid: "b3a88af1-3568-4219-99c4-aa8fbce3227d",
+    name: "ANUGA",
+    description:
+      "ANUGA模型是一种流体力学建模工具，允许用户在复杂的二维几何形状中建模模拟现实的流动问题。是由Stephen Roberts，Ole Nielsen,Duncan Gray,Jane Sexton等人开发。ANUGA的核心是流体动力学模块，称为Shallow_water，它是基于求解浅水波方程的有限体积方法。研究区域用三角形网格表示。通过求解每个单元内的控制方程，可以跟踪水深和水平动量随时间的变化。",
+    author: "Stephen Roberts，Ole Nielsen,Duncan Gray,Jane Sexton",
+    select:false,
+    output:'output_anuga',
+  }        
+];
+  
+let output = [{
+    label:'1',
+    path:'',
+    value:'report1.rpt'
+  },{
+    label:'2',
+    path:'',
+    value:'output1.out'
+  },{
+    label:'3',
+    path:'res_clip_testNoBciAndDoubleCouple1',
+    value:'fenhu-0000.wd',
+  },{
+    label:'4',
+    path:'res_clip_testNoBciAndDoubleCouple1',
+    value:'fenhu-0001.wd',
+  },{
+    label:'5',
+    path:'res_clip_testNoBciAndDoubleCouple1',
+    value:'fenhu-0002.wd',
   },
-  mounted() {
-
-  },
-  computed:{
-    selected(){
-      return this.value;
-    }
+];
+    
+//#region 
+//关闭dialogVisible对话框
+const handleClose = (done) => {
+  dialog.dialogVisible = false;
+};
+//关闭deployDocument对话框
+const handleCloseDeployDocument = (done) => {
+  dialog.deployDocument = false;
+};
+//搜索模型
+const searchModelItem = () => {};
+//选择需要耦合的模型
+const selectModelItem = (index, info)  =>{
+  let modelObj = {
+    oid:info.oid,
+    name:info.name,
+    description:info.description,
+    author:info.author,
+  }
+  modelData.modelList.push(modelObj);
+  ElMessage({
+    message:'选择模型成功',
+    type:'success',
+  })
+  info.select = !info.select;
+};
+//载入已选的模型
+const loadModelItem = (index, info)  =>{
+  xml_show.value = '';
+  dialog.deployDocument = true;
+  let modelObj = {
+    oid:info.oid,
+    name:info.name,
+    description:info.description,
+    author:info.author,
+  }
+  modelData.loadModel = modelObj;
+  modelData.loadModels.push(modelObj);
+};
+//调用上传的删除方法，清除上传文件列表
+const clearFiles = () => {
+  if(configUpload.value !== undefined){     
+    configUpload.value.clearFiles();
   }
 };
+//上传文件，并展示内容到页面上
+const uploadChange = (file, fileList)  =>{
+  if (file.status !== 'ready') return;
+  let fileObj = {
+    file:fileList,
+    name:modelData.loadModel.name,
+    description:modelData.loadModel.description,
+    author:modelData.loadModel.author,        
+    oid:modelData.loadModel.oid,
+  }
+  modelData.uploadFiles.push(fileObj);
+  
+  let form = new FormData();
+  for (let i = 0; i < modelData.uploadFiles.length; i++) {
+    //上传文件信息以及模型信息
+    form.append("datafile", modelData.uploadFiles[i].raw);
+  }
+
+  for(let item of modelData.uploadFiles){
+    //上传文件信息以及模型信息
+    form.append("datafile", item.file[0].raw);
+    form.append("name", item.name);
+    form.append("description", item.description);
+    form.append("author", item.author);
+    form.append("oid", item.oid);
+  }
+  axios.post("/api/coupleDocument", form).then(res => {
+    //解析该文件，返回xml字符串，赋值给xml_show
+    if (res.status === 200) {
+      let result = res.data;
+      if (result.code === 0) {
+        ElMessage({
+          message: "载入成功",
+          type: "success"
+        });
+        xml_show.value = result.data;
+      } else {
+        ElMessage.error("载入失败");
+      }
+    } else {
+      ElMessage.error("服务器error");
+    }
+  });
+};
+//载入配置文件到模型模块化耦合控制区操作面板
+const confirmLoad = () => {    
+  modelData.modelConfig.push({
+    name: modelData.loadModel.name,
+    oid:modelData.loadModel.oid,
+    configName: modelData.loadModel.name + ".conf",
+  });
+  dialog.deployDocument = false;
+};
+//调用函数
+const invoke = ()  =>{
+  //根据文件oid，调用exe，生成最终结果，并可视化
+  let form = new FormData();
+  let oid = '', oids = [];
+  for(let item of modelData.loadModels){
+    if(oid !== item.oid){
+      oid = item.oid;
+      oids.push(oid);
+    }
+  }
+  if(oids.length < 2){
+    ElMessage({
+      message:"模型数目不够2个，无法耦合",
+      type:'error',
+    })
+  }else{
+    form.append('oids', oids);
+    load.loading = true;
+    setTimeout(() => {
+      isShow.value = true;
+      load.loading = false;
+    }, 1000)
+  }
+};
+//配置文件下载函数
+const downloadConfig = (oid) =>{
+  console.log(`oid 为 ${oid}`);
+  if(oid === undefined){
+    ElMessage({message:'该配置文件无oid，请检查！', type:'error'});
+  }
+  let form = new FormData();
+  form.append('oid',oid);
+  load.downloadLoading = true;
+  
+  window.location.href = `/api/downloadConfig?oid=${oid}`;  
+  load.downloadLoading = false;
+};
+//下载输出数据
+const downloadOutput = () =>{
+  if(selected.value === ''){
+    ElMessage({message:"未选择需要下载的文件，请先选择待下载文件！", type:"error"})
+    return;
+  }
+  //拼写出下载的文件路径
+  let path = '';
+  for(let item of output){
+    if(item.value === selected.value){
+      if(item.path === ''){
+        path = outputPath.value + '/' + selected.value;
+      }else{
+        path = outputPath.value + '/' + item.path + '/' + selected.value;
+      }          
+      break;
+    }
+  }
+  let form = new FormData();
+  form.append('filePath', path);
+  load.downloadFileLoading = true;
+  //利用blob实现axios下载,超时时间五分钟
+  axios.post('/api/downloadFile', form, {responseType:'blob'},{timeout:30000}).then(res => {
+    load.downloadFileLoading = false;
+
+    const content = res.data;
+    const blob = new Blob([content]);
+    if('download' in document.createElement('a')){
+      //非IE下载
+      const elink = document.createElement('a');
+      elink.download = selected.value;
+      elink.style.display = 'none';
+      elink.href = window.URL.createObjectURL(blob);
+      document.body.appendChild(elink);
+      elink.click();
+      window.URL.revokeObjectURL(elink.href);
+      document.body.removeChild(elink);
+    }else{
+      //IE10+下载
+      if(typeof window.navigator.msSaveBlob !== 'undefined'){
+        window.navigator.msSaveBlob(blob, selected.value);
+      }else{
+        let URL = window.URL || window.webkitURL;
+        let downloadUrl = URL.createObjectURL(blob);
+        window.location = downloadUrl;
+      }
+    }
+  }).catch(err => {
+    console.log(err);
+  })      
+};
+
+const showMap = () => {
+  router.push('/couple')
+}
+//#endregion
+
+let selected = computed({
+    get(){
+        return value.value;
+    },
+    set(value){
+        value.value = value;
+    }
+})
 </script>
-<style scoped>
+<style>
 .parent {
   border: 1px solid #222;
   display: flex; /*设为伸缩容器*/
